# swagger/swagger.yaml
swagger: "2.0"
info:
  title: Identity Manager
  description: |
    This is the base API for Identity Manager microservice. Api documentation
    available in /api/idm/docs
  version: "1.0.0"
  contact:
    name: Carlos Ruiz
    email: caruizag@gmail.com
host: localhost:30030
basePath: /api/idm
schemes:
  - http
paths:
...
  /resource/{id}:
    parameters:
      - name: id
        in: path
        description: Login ID to get
        type: string
        required: true
    get:
      tags:
        - resource
      summary: Get Resource details by Id
      description: |
        Get a particular resource by Id
      operationId: getResourceById
      produces:
        - application/json
      security:
        - bearer: []
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/Resource'
        404:
          description: Resource not found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /resource/:
    put:
      tags:
        - resource
      summary: Update Resource details by Id
      description: |
        Get a particular resource by Id
      operationId: updateResource
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - bearer: []
      parameters:
        - in: body
          name: resource
          description: resource
          required: true
          schema:
            $ref: '#/definitions/Resource'
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/Resource'
        404:
          description: Resource not found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      tags:
        - resource
      summary: Add Resource details by Id
      description: |
        Get a particular resource by Id
      operationId: addResource
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - bearer: []
      parameters:
        - in: body
          name: resource
          required: true
          schema:
            $ref: '#/definitions/Resource'
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/Resource'
        404:
          description: Resource not found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - resource
      summary: Delete Resource details by Id
      description: |
        Get a particular resource by Id
      operationId: deleteResource
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - bearer: []
      parameters:
        - in: body
          name: resource
          required: true
          schema:
            $ref: '#/definitions/Resource'
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/Resource'
        404:
          description: Resource not found
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
securityDefinitions:
  bearer:
     type: apiKey
     in: header
     name: Authorization
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
    required:
      - message
  Resource:
    type: object
    properties:
      id:
        type: string
      description:
        type: string
    required:
      - id
...